Cleanup/restructure backend
