+++ src/auth/middleware.rs
@@ -1,4 +1,4 @@
-use crate::ctx::Ctx;
+use ctx::Ctx;
 use crate::auth::{COOKIE_NAME, KEY};
 use crate::{Error, Result};
 use axum::{
