+++ src/auth/mod.rs
@@ -4,6 +4,7 @@ use tower_cookies::Key;
 
 pub mod middleware;
 pub mod routes;
+pub mod ctx;
 
 pub const COOKIE_NAME: &str = "AUTH_TOKEN";
